release 1.2
Some minor bug fix, updated readme and forgot-password addded

+++ config/config.go
@@ -74,7 +74,7 @@ type Redeem struct {
 
 type Reward struct {
 	Time	string `json:"time"`
-	Coins	int    `json:"coins"`
+	Coins	float64    `json:"coins"`
 	Remarks string `json:"remarks"`
 }
